+++ src/common/_selectors.scss
@@ -22,3 +22,7 @@
   @extend .text-small;
   font-weight: 700;
 }
+
+.text-error-message {
+  @include text(13px, 400, 16px);
+}
