+++ src/stories/forms/InputComment.stories.ts
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
-import { InputComment } from './InputComment';
+import { InputComment } from '../../ui/forms/InputComment';
 
 // More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction
 const meta = {
