+++ src/stories/forms/InputName.stories.ts
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
-import { InputName } from './InputName';
+import { InputName } from '../../ui/forms/InputName';
 
 // More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction
 const meta = {
