+++ src/stories/forms/InputPhoneNumber.stories.ts
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
-import { InputPhoneNumber } from './InputPhoneNumber';
+import { InputPhoneNumber } from '../../ui/forms/InputPhoneNumber';
 
 // More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction
 const meta = {
