+++ src/stories/forms/SelectSchool.stories.ts
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
-import { SelectSchool } from './SelectSchool';
+import { SelectSchool } from '../../ui/forms/SelectSchool';
 
 // More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction
 const meta = {
