+++ src/ui/forms/SelectSchool.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import styles from './input.module.scss';
+import styles from './selectSchool.module.scss';
 
 interface Props {
   /**
