+++ src/ui/forms/input.module.scss
@@ -52,7 +52,7 @@
   @extend .text-medium;
   letter-spacing: -0.32px;
   appearance: none;
-  background-image: url('../assets/dropdown.svg');
+  background-image: url('../../stories/assets/dropdown.svg');
   background-repeat: no-repeat;
   background-position: right 0.7em top 50%, 0 0;
   background-size: 14px, 7px;
