+++ src/ui/forms/inputName.module.scss
@@ -1,0 +1,45 @@
+@import '../../common/_selectors';
+@import '../../common/_variable';
+
+.form {
+  width: 375px;
+  height: 96px;
+  display: flex;
+  flex-direction: column;
+  justify-content: start;
+  background: map-get($background_color, 'primary');
+  padding: 12px;
+}
+
+.label {
+  width: 351px;
+  height: 18px;
+  font-family: 'GolosText';
+  font-style: normal;
+  font-weight: 400;
+  font-size: 14px;
+  line-height: 18px;
+  letter-spacing: -0.154px;
+  color: map-get($background_color, 'partsPrimary');
+  margin-bottom: 8px;
+}
+
+.input {
+  width: 351px;
+  height: 44px;
+  padding: 12px;
+  box-sizing: border-box;
+  background: map-get($background_color, 'primary');
+  border: 1px solid map-get($background_color, 'partsSecondary');
+  @extend .text-medium;
+  letter-spacing: -0.32px;
+  color: map-get($background_color, 'partsPrimary');
+}
+
+
+.input:focus-visible,
+.input:focus {
+  outline: 1px solid map-get($background_color, 'partsActiveSecondary');
+  border: 1px solid map-get($background_color, 'partsActiveSecondary');
+}
+
